Add go.mod file
This adds a go.mod file, to better support managing dependencies in Go
with modules. All of the cool kids started doing this with the
preliminary support in Go v1.11 and v1.12, and it became the default
with Go 1.13, released in September 2019.

If you clone the repository and try to run the benchmark with the latest
Go 1.17 compiler, you get an error:

```
$ go test -v -bench=.
go: cannot find main module, but found .git/config in ~/src/github.com/mojotx/pcg
	to create a module there, run:
	go mod init
```